move azuretest package to azure functional tests package
