+++ docs/contributing/contributing-code/contributing-code-tests/writing-functional-tests.md
@@ -17,24 +17,26 @@ For example a test for deploying an Azure Service Bus resource would be in `/tes
 Tests should look like the following. You can actually copy-paste this to create a new test!
 
 ```go
-func Test_DescriptiveTestName(t *testing.T) {
-	application := "unique-application-name"
-	template := "testdata/unique-application-name.bicep"
-	test := azuretest.NewApplicationTest(t, application, []azuretest.Step{
-		{
-			Executor: executor.NewDeployStepExecutor(template, ""),
-            Components: &validation.ComponentSet{
-                // Set of components to validate
-            },
-			Pods: &validation.K8sObjectSet{
-				// set of K8s resources to validate
-			},
-            // This should be set to true for every test for now
-			SkipARMResources: true,
-		},
-	})
+import "github.com/project-radius/radius/test/functional/azure"
 
-	test.Test(t)
+func Test_DescriptiveTestName(t *testing.T) {
+    application := "unique-application-name"
+    template := "testdata/unique-application-name.bicep"
+    test := azure.NewApplicationTest(t, application, []azure.Step{
+        {
+            Executor: executor.NewDeployStepExecutor(template, ""),
+            Components: &validation.ComponentSet{
+            // Set of components to validate
+        },
+        Pods: &validation.K8sObjectSet{
+        // set of K8s resources to validate
+        },
+        // This should be set to true for every test for now
+        SkipARMResources: true,
+        },
+    })
+
+    test.Test(t)
 }
 ```
 
@@ -44,22 +46,24 @@ When adding a new functional test:
 - Double-check that the application name is unique (do a search in the repo)
 - Avoid skipping any verifications (other than `SkipARMResources`)
 - Avoid using `PostStepVerify` and `PostDeleteVerify` if you can add new capabilities to the test system
-- For the tests to verify that the containers are actually started and in Ready state, you can add a readiness probe to the bicep file as below. 
-	```bicep
-	resource a 'Container' = {
-		name: 'a'
-		properties: {
-		container: {
-			image: '${registry}/magpiego:latest' // This image implements readiness checks
-			env: {
-			COOL_SETTING: env
-			}
-			readinessProbe:{
-			kind:'httpGet'
-			containerPort:3000
-			path: '/healthz'
-			}
-		}
-		}
-	}
-	```
+- For the tests to verify that the containers are actually started and in Ready state, you can add a readiness probe to the bicep file as below.
+
+    ```bicep
+    resource a 'Container' = {
+        name: 'a'
+        properties: {
+            container: {
+                // This image implements readiness checks
+                image: '${registry}/magpiego:latest' 
+                env: {
+                    COOL_SETTING: env
+                }
+                readinessProbe:{
+                    kind:'httpGet'
+                    containerPort:3000
+                    path: '/healthz'
+                }
+            }
+        }
+    }
+    ```
