+++ test/functional/azure/applicationtest.go
@@ -3,7 +3,7 @@
 // Licensed under the MIT License.
 // ------------------------------------------------------------
 
-package azuretest
+package azure
 
 import (
 	"context"
