+++ test/functional/azure/azuretest.go
@@ -3,7 +3,7 @@
 // Licensed under the MIT License.
 // ------------------------------------------------------------
 
-package azuretest
+package azure
 
 import (
 	"testing"
