+++ test/functional/azure/cli/appdeploy_test.go
@@ -8,18 +8,19 @@ package cli_test
 import (
 	"testing"
 
-	"github.com/project-radius/radius/test/azuretest"
+	"github.com/stretchr/testify/require"
+
+	"github.com/project-radius/radius/test/functional/azure"
 	"github.com/project-radius/radius/test/radcli"
 	"github.com/project-radius/radius/test/testcontext"
 	"github.com/project-radius/radius/test/validation"
-	"github.com/stretchr/testify/require"
 )
 
 func Test_AppDeploy_ScaffoldedApp(t *testing.T) {
 	ctx, cancel := testcontext.GetContext(t)
 	defer cancel()
 
-	options := azuretest.NewTestOptions(t)
+	options := azure.NewTestOptions(t)
 
 	// We deploy a simple app and then run a variety of different CLI commands on it. Emphasis here
 	// is on the commands that aren't tested as part of our main flow.
