+++ test/functional/azure/mechanics/mechanics_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/project-radius/radius/pkg/renderers/containerv1alpha3"
 	"github.com/project-radius/radius/pkg/renderers/httproutev1alpha3"
 	"github.com/project-radius/radius/pkg/resourcekinds"
-	"github.com/project-radius/radius/test/azuretest"
+	"github.com/project-radius/radius/test/functional/azure"
 	"github.com/project-radius/radius/test/executor"
 	"github.com/project-radius/radius/test/functional"
 	"github.com/project-radius/radius/test/validation"
@@ -27,7 +27,7 @@ func Test_RedeployWithAnotherResource(t *testing.T) {
 	application := "azure-mechanics-redeploy-withanotherresource"
 	templateFmt := "testdata/azure-mechanics-redeploy-withanotherresource.step%d.bicep"
 
-	test := azuretest.NewApplicationTest(t, application, []azuretest.Step{
+	test := azure.NewApplicationTest(t, application, []azure.Step{
 		{
 			Executor: executor.NewDeployStepExecutor(fmt.Sprintf(templateFmt, 1), functional.GetMagpieImage()),
 			AzureResources: &validation.AzureResourceSet{
@@ -100,7 +100,7 @@ func Test_CommunicationCycle(t *testing.T) {
 	application := "azure-mechanics-communication-cycle"
 	template := "testdata/azure-mechanics-communication-cycle.bicep"
 
-	test := azuretest.NewApplicationTest(t, application, []azuretest.Step{
+	test := azure.NewApplicationTest(t, application, []azure.Step{
 		{
 			Executor: executor.NewDeployStepExecutor(template, functional.GetMagpieImage()),
 			AzureResources: &validation.AzureResourceSet{
