+++ test/functional/azure/resources/dapr_pubsub_servicebus_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/project-radius/radius/pkg/renderers/containerv1alpha3"
 	"github.com/project-radius/radius/pkg/renderers/daprpubsubv1alpha3"
 	"github.com/project-radius/radius/pkg/resourcekinds"
-	"github.com/project-radius/radius/test/azuretest"
+	"github.com/project-radius/radius/test/functional/azure"
 	"github.com/project-radius/radius/test/executor"
 	"github.com/project-radius/radius/test/functional"
 	"github.com/project-radius/radius/test/validation"
@@ -25,7 +25,7 @@ func Test_DaprPubSubServiceBus(t *testing.T) {
 	application := "azure-resources-dapr-pubsub-servicebus"
 	template := "testdata/azure-resources-dapr-pubsub-servicebus.bicep"
 
-	test := azuretest.NewApplicationTest(t, application, []azuretest.Step{
+	test := azure.NewApplicationTest(t, application, []azure.Step{
 		{
 			Executor: executor.NewDeployStepExecutor(template, functional.GetMagpieImage()),
 			AzureResources: &validation.AzureResourceSet{
