+++ test/functional/azure/resources/dapr_statestore_tablestorage_test.go
@@ -14,9 +14,9 @@ import (
 	"github.com/project-radius/radius/pkg/renderers/containerv1alpha3"
 	"github.com/project-radius/radius/pkg/renderers/daprstatestorev1alpha3"
 	"github.com/project-radius/radius/pkg/resourcekinds"
-	"github.com/project-radius/radius/test/azuretest"
 	"github.com/project-radius/radius/test/executor"
 	"github.com/project-radius/radius/test/functional"
+	"github.com/project-radius/radius/test/functional/azure"
 	"github.com/project-radius/radius/test/validation"
 )
 
@@ -24,7 +24,7 @@ func Test_DaprStateStoreTableStorage(t *testing.T) {
 	application := "azure-resources-dapr-statestore-tablestorage"
 	template := "testdata/azure-resources-dapr-statestore-tablestorage.bicep"
 
-	test := azuretest.NewApplicationTest(t, application, []azuretest.Step{
+	test := azure.NewApplicationTest(t, application, []azure.Step{
 		{
 			Executor:           executor.NewDeployStepExecutor(template, functional.GetMagpieImage()),
 			AzureResources:     &validation.AzureResourceSet{},
